+++ src/components/Comments.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react"
-
+import "../styles/comments.css";
