+++ src/styles/discussion.css
@@ -1,0 +1,126 @@
+
+.discussionMain-container{
+    text-align: center;
+}
+body {
+    --off-white: whitesmoke;
+    --primary-color: rgb(231, 143, 28);
+    --primary-text-color: rgb(223, 223, 223);
+    --line-color: rgb(59, 58, 58);
+    margin: 0 10em;
+  }
+  
+  .main-container {
+   display: flex;
+   flex-direction: row;
+    /* display: grid;
+    grid-template-columns: 1fr 4fr; */
+   /*  grid-template-rows: 4fr 1fr; */
+   padding: 2em 10em;
+   background: var(--primary-color);
+   margin: 0 5em;
+  }
+  main section {
+  text-align: left;
+  padding: .5em;
+  }
+  
+  .friends {
+    background: rgb(249, 167, 39);
+    grid-column: 1 / span 1;
+    grid-row: 1 / span 2;
+    /* flex-grow: 1; */
+    word-wrap: nowrap;
+    
+  }
+  .friends ul {
+    margin-block-start: 0;
+    margin-block-end: 0;
+  }
+  
+  .messages {
+    background: var(--primary-color);
+    grid-column: 2 / span 1;
+    grid-row: 1 /span 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  .messages .header {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+  }
+  
+  .messages h4 {
+    font-size: .6em;
+  }
+  
+  .messages article {
+    padding: 1em;
+    color: var(--primary-text-color);
+    margin: 0 0 .5em 0;
+    border-bottom: 1px solid var(--line-color);
+    width: 100%;
+   
+  }
+  
+  .messages .new-message {
+    grid-column: 2 /span 1 ;
+    grid-row: 2 /span 1;
+    border-bottom: none;
+  }
+  
+  .new-message form {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .new-message form button {
+    background: white;
+    border-radius: 1.5em;
+    padding: .5em 1em;
+  
+  }
+  
+  .new-message form .input-name {
+    border-radius: 2em;
+    border: none;
+    padding: .5em 1em;
+  }
+  
+  .new-message form .input-text {
+    border-radius: 1em;
+    border: none;
+    padding: 1em;
+    min-height: 7em;
+    margin: 1em 0;
+  }
+  
+  li {
+    list-style: none;
+  }
+  
+  
+  
+  
+  .reaction-container {
+    display: flex;
+    flex-direction: row;
+    color: var(--primary-text-color);
+    justify-content: left;
+    padding: .5em 0;
+  }
+  
+  .reaction-container button{
+    background: none;
+    border: none;
+    font-size: .8em;
+    padding: 0;
+  }
+  
+   .active-post{
+    background-color: darkgreen;
+    color: rgb(214, 207, 207);
+  }
